# src/main/resources/swagger/ex-website-api.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "ex-website-api",
      "contact": {
        "email": "maestrobo1@mail.ru"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "Generated server url"
      }
    ],
    "paths": {
      "/blog/{id}/edit": {
        "get": {
          "tags": [
            "blog-page-controller"
          ],
          "operationId": "blogEdit",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "blog-page-controller"
          ],
          "operationId": "postBlogUpdate",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "title",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "topic",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "fullText",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/registration/register": {
        "post": {
          "tags": [
            "reg-controller"
          ],
          "operationId": "postUser",
          "parameters": [
            {
              "name": "regDto",
              "in": "query",
              "required": true,
              "schema": {
                "$ref": "#/components/schemas/RegDto"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/blog/post": {
        "get": {
          "tags": [
            "blog-page-controller"
          ],
          "operationId": "addPost",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "blog-page-controller"
          ],
          "operationId": "postBlog",
          "parameters": [
            {
              "name": "title",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "topic",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "fullText",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/registration/open": {
        "get": {
          "tags": [
            "reg-controller"
          ],
          "operationId": "openRegPage",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/blog": {
        "get": {
          "tags": [
            "blog-page-controller"
          ],
          "operationId": "allBlogs",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/blog/{id}": {
        "get": {
          "tags": [
            "blog-page-controller"
          ],
          "operationId": "blogDetail",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/auth": {
        "get": {
          "tags": [
            "blog-page-controller"
          ],
          "operationId": "auth",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/about": {
        "get": {
          "tags": [
            "blog-page-controller"
          ],
          "operationId": "about",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/": {
        "get": {
          "tags": [
            "blog-page-controller"
          ],
          "operationId": "home",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/blog/{id}/remove": {
        "delete": {
          "tags": [
            "blog-page-controller"
          ],
          "operationId": "postBlogDelete",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "RegDto": {
          "type": "object",
          "properties": {
            "login": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "confirm": {
              "type": "string"
            }
          }
        }
      }
    }
  }